refactor: query-key-factory 컨벤션에 맞게 수정 (#118)
