+++ apps/web/src/entities/auction/queryKeyFactory.ts
