+++ apps/web/src/features/auction/AuctionFormPage.tsx
@@ -4,7 +4,7 @@
 
 import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
 import { getAddressId, getAuction } from 'src/entities/auction/api';
-import { addressIdKeys, auctionFormKeys } from 'src/entities/auction/query-key-factory';
+import { addressIdKeys, auctionFormKeys } from 'src/entities/auction/queryKeyFactory';
 import AuctionForm from 'src/features/auction/AuctionForm';
 import DetailPageHeader from 'src/widgets/DetailPageHeader';
